Add files via upload
Data --> https://www.kaggle.com/cdc/behavioral-risk-factor-surveillance-system

Summary --> To put into context where we stand in the world, we are currently at the top of the chart when it comes to the obesity percent of the population. In Figure 3 below, one can see that the US is 1st with Mexico 2nd. Now as stated before, people have different opinions as to why we have a rising obesity issue in the US. Our group believes that there is a link between socioeconomic status, and current weight. We believe these factors play a significant role in the current epidemic, as it will be stated in the hypothesis. Now as we move on, we feel it is important to reiterate that it is unsustainable for a country to ignore or barely do anything about the growing epidemic as there are significant consequences to a growing population, over time there will be a major health and social issues regarding this so now is the best time to address the roots of the problem. In our data, which will later be introduced, the average age is 54 see Figure 4. Meaning this follows closely with the CDC data on obesity.